Refactor authentication route
